Refactored code to include custom hooks for selected options and correctness to improve readability and reusability
